requirements.txt
fastapi
python-multipart
jinja2
Pillow
tensorflow
numpy
aiofiles
pydantic